Re-enable SafeArgs java plugin tests
There were disabled when Androidx was moving to gradle 7.0 and it seems
that the blocking issues have been resolved.

There are a couple of issues that this change does work around:

1. It seems like to use `ext` variables defined in the buildscript block
directly in the module's android block, the variables now have to be
full qualified by using the rootProject. (b/203721113)
2. Playground settings are causing the to pick the wrong prebuilts
directory. (b/203721036)

Test: ./gradlew --rerun-tasks navigation:navigation-safe-args-gradle-plugin:test
Bug: 186177309
Change-Id: I66d288c18d37582b11041dcc14f6e585fb40f6da