Improve onNavDestinationSelected return logic
When the listeners for NavigationBarView and
NavigationView return true, they automatically
set that MenuItem to selected. This automatic
selection happens *after* the
OnDestinationChangedListener fires from the
navigate() call. This meant that the selected
MenuItem would fall out of sync with the current
navigation graph in a few key spots:

1) When you navigate to an <activity> destination -
in this case, the navigate() completes successfully,
but we don't want to update the MenuItem to be
selected (as you aren't actually on that destination)
2) When you've mixed destinations from another navigation
graph onto the back stack associated with another
graph - in this case, the destination you end up on
may not be associated with the MenuItem you've selected.

In each case, the return value of
onNavDestinationSelected should match the logic in each
OnDestinationChangedListener and use matchDestination()
to determine if you've actually navigated to that graph.

Test: tested in integration-tests app, sample app in bug
BUG: 160478847
Relnote: "Fixed an issue where the
`onNavDestinationSelected` API on `NavigationUI` would
return `true` even if you didn't actually `navigate()` to
that navigation graph. It now uses the same logic as is
used internally in `setupWithNavController` to only
select the `MenuItem` associated with the current
destination using the `hierarchy` of the destination."

Change-Id: I2b053225195ac72ee7e2e6aabebcbb0b97418ec0