Fix stretched preview in CoreTest-OpenGL
The streched preview happens in CameraX core test
-OpenGL activity Or core test app with the modification
to remove UseCaseGroup.

Cause:
When UseCaseGroup is not used, it will need to calcualte
the mPreviewCropRect from getTextureRotationDegrees and
SurfaceRotationDegrees.  However,  when getTextureRotationDegrees
changes, it checks if mPreviewCropRect is null
to see if need to update mPreviewCropRect but mPreviewCropRect will
be non-null after renderLatest is called.

Bug: NA
Test: Manual test.
Change-Id: I83d175408dd6ec57ebe0320dcdb2b8f4074c8ef9