[GH] Avoid duplicate project dirs in playground
This CL fixes a bug in playground where we might end up with duplicate
project directories when Gradle automatically creates a parent project
when it does not exist. When that case happens, Gradle will pick a
project directory relative to the playground root project. Instead, we
provide a project directory relative to the child project to ensure
consistency.

With this change, we can also enable to the project path validation in
the AndroidXPlugin.

Bug: 197253160
Test: cd activity && ./gradlew tasks // fails without this change

This is an imported pull request from https://github.com/androidx/androidx/pull/260.

Resolves #260
Github-Pr-Head-Sha: b6ddf646a8289b52b725264807b288a9a48cdf77
GitOrigin-RevId: 0c4d385cb44c772508d9b5e7e4561f268be83785
Change-Id: I5b4ac8104172bfae5c87cb725fa7e2fd38aac54e