LazyList item click handling
Support click handling with LazyColumn by creating a list adapter
PendingIntent template, and filling in for each click action.

For action callbacks, we use an activity trampoline to begin the
callback broadcast receiver.

For activity starts, to properly fill in the pending intent in such a
way that they are not filter equals, we either set an identifier on the
activity intent after Api 29, or use an activity trampoline with a
unique data field before Api 29.

Bug: 198618359
Test: ./gradlew :glance:glance-appwidget:connectedAndroidTest
Test: ./gradlew :glance:glance-appwidget:integration-tests:demos:installDebug
Change-Id: I72e7e5e5538750383bd2745cda3898f0d10503aa