Add code to clean the list of receivers on update
If the package get updated, receivers may be removed, at which point we
need to remove them from our database too. For those whose association
changes, they will be updated either soon if App Widgets already exist,
or as soon as App Widget will be added. And no API will be incorrect in
the meantime.

Bug: 205957419
Test: demos
Test: ./gradlew :glance:glance-appwidget:connectedAndroidTest
Relnote: n/a
Change-Id: Id85d4b64aaccccaae19ff482638f4237ccb7c3a6