Fix Activity leak in Navigation compose
Navigation currently does not set the SaveableStateHolder for entries
that are not on top of the back stack when the NavController is restored
due to a config change or process death. This means that these entries
hold on to old instances of saveableStateHolder, which means they
maintain a reference to the activity and cause a leak since the old
activity has already been destroyed.

We need to clear the saveableStateHolder of entries on the back stack
immediately once we dispose of the current destination.

RelNote: "Fixed a leak in Navigation Compose where it was holding on to
a reference of the old activity after a config change or process death."
Test: NavHostTest
Bug: 204905432

Change-Id: I4efcb0709b3864dde243d011548d801fcef24159