Fix intermittent crash in ModifierLocal snapshot observer
From the logs, it looks like the
error is caused when a modifier local provider's value
changes when the layout node associated with it is not
attached.

This CL just ignores the value change when the layout node
is not attached, and does not notify the consumers of that
value. We can safely do this because whenever the modifier
attached it will send a notification to all the consumers.

Bug: 205414040
Fixes: 205414040
Test: ModifierLocalConsumerNodeTest
Change-Id: I61a95f612bac0bd12558a7aa9b742a709d1885a6