[GH] Fix cacheability of GenerateAntlrGrammarTask
## Proposed Changes

By extending `JavaExec`, the `generateAntlrGrammarTask` declares that the `args` parameter is a task input. Since this value contains the absolute path of an input file, this can be the source of cache misses.

This change converts the task to a `DefaultTask` and uses composition to delegate work to `ExecOperations.javaexec`. This allows the task to fully declare it's inputs and outputs in a path-sensitive manner.

## Testing

Test: [Running the same build from different checkout locations via Github Actions](https://github.com/bigdaz/androidx/blob/ge-experiments/.github/workflows/ge-experiments.yml) was used to demonstrate the initial cache miss, which is fixed by this PR.

This is an imported pull request from https://github.com/androidx/androidx/pull/282.

Resolves #282
Github-Pr-Head-Sha: db918f920fcae5928b92613f15a12866608c829f
GitOrigin-RevId: 936f9172dba6748689ca22ae0b865c5f41996d48
Change-Id: I0610d2a5421f9646efdbf6dcbb4e3a87772555c8