Attach Fragment's SavedStateRegistry in PreAttach
By using SavedStateRegistryController's new
performAttach() method, we can ensure that the
LifecycleObserver added by that method is added
to the Fragment as part of its pre attach listeners.

This ensures that its LifecycleObserver runs before
any LifecycleObserver added by
registerForActivityResult().

Relnote: "Fixed a crash when accessing a `ViewModel`
for the very first time from a
`registerForActivityResult()` callback of a
`Fragment`."
Test: savedStateOnActivityResult
BUG: 177246258

Change-Id: Iea2b30ca03908798474cac9327c4e06e3d4fd09b