Refactored API to clean up singleton usage
Various changes according to ongoing API reviews. The major
change is a refactor which splits out the state logic of JankStats
into the new class ApplicationState.

PerformanceMetricsState is a per-window singleton (as JankStats
was previously) which
allows storage and retrieval of app state in a single object within
the hierarchy. Meanwhile, JankStats is now a regular object (not a
singleton), created via constructor, that is wholly managed by the
application. This division should make it clearer and cleaner such
that logic outside of the app code (e.g., RecyclerView) can retrieve
the PerformanceMetricsState and store information there without needing to
deal with (and potentially interfere with) JankStats.

Note that PerformanceMetricsState could even be used by other non-JankStats
libraries eventually; it is not specific to JankStats (though it still
contains frame timing logic, so it is best used by performance clients).

Also cleaned up other APIs and functionality:

Removed clearStats().
Documented example usage of jankHeuristicMultipler.
FrameMetrics usaged protected against doubled reports.
JankStats now takes a Window directly. PerformanceMetricsState takes a view
in the hierarchy, similar to the previous JankStats singleton.
PerformanceMetricsState is cached inside a MetricsStateHolder object.
Clients should request a holder and can then query it quickly to
get the state, rather than having to call the instance getter each time,
which walks the view hierarchy up to the root.

Bug: 205904685
Bug: 205806861
Test: Manual (integration tests and JankStats runs across devices)
Relnote: Removed clearStates() based on API review
Change-Id: I489696a1383a09be1222d5bb744a3615ba176c56