Avoid saving an empty Bundle for SavedStateRegistry
When performSave() is called, SavedStateRegistry
might actually not have anything to save. In those
cases, it should avoid adding an empty Bundle to
the outBundle.

This allows systems that use SavedStateRegistry
(such as Fragments) to use Bundle.isEmpty() as a
sign on whether it needs to add its Bundle, etc.

Test: new SavedStateRegistryTest test
BUG: 203457956
Change-Id: Ib49bd8645a5d4e38de329e50fe659b1baf7fcba3