Expose watch face ID via WatchState
Test: Manual testing
Bug: 192981547
Relnote: The watch face's instance ID is now available via WatchState.watchFaceInstanceId. Most watch faces won't need to use this, but if there's per-watch face state that's not stored in the Schema then this is the key to use to identify the watch face instance. To help support this you can now provide an ID when calling WatchFaceControlClient.createHeadlessWatchFaceClient.
Change-Id: I1ff987d6920b71b736bd3e4c3ac8c2f4b658b77b