Consolidate behaviour of receiver parameter between KAPT and KSP
This change adds a synthetic receiver parameter to KSP method elements that are extension functions. Similar to suspend functions the synthesised element makes it so that KAPT and KSP parameters are the same. This also fixes an issue in Javac / KAPT where the receiver parameter name was returning the next parameter name from the Kotlin metadata because metadata does not contain the receiver and there was an off by one error.

Bug: 160322705
Bug: 207676234
Test: XExecutableElementTest
Change-Id: I1b71df1ed6ae7dd14d012e0c12d9b28043a7847e