Extend ScalingLazyColumn to support scrollTo and animatedScrollTo.
Expose the scrollTo/animatedScrollTo methods to the ScalingLazyColumn class. In order to do this we need to make the ScalingLazyColumn and ScalingLazyListState consider the position of items with respect to the center of the viewport and not the top. This means that scrollTo() will move the selected item to be at the center of the viewport, taking a ScalingLazyListAnchorType (ItemCenter|ItemStart) and offset into account.

Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Bug: 208654683
Relnote: "We have added a number of methods to ScalingLazyListState to allow developer control over scrolling to specific list items and also setting the initial list item and offset. The new methods are scrollTo, animatedScrollTo, centerItemIndex and centerItemOffset."
Change-Id: I61b61fbed9539408138216fb9acef46c9bd94479