Don't wrap NumberKeyListener for emoji2
This avoids wrapping instances of NumberKeyListener when passed via
setKeyListener. TextView will depend on this type to do locale
configuration of punctionation, and these keylisteners never allow
inputting emoji.

Relnote: "Emoji2 will not wrap instances of NumberKeyListener, allowing
the the locale to be configured by textview."
Relnote: "Appcompat will not wrap instances of NumberKeyListener passed
to setKeyListener, allowing TextView to correctly configure the locale
on NumberKeyListeners."

Fixes: b/207119921
Test: ./gradlew :e2:e2:cAT :e2:e2-bun:cAT :appc:appc:cAT
Change-Id: Ibf113081112d05c75937eec5bc87904ebf450b26