Fix re-adding fragments removed before initial attach
When a Fragment is removed from the FragmentManager
before it even goes through onAttach(), we have
special code that runs to clean up the Fragment.

However, initLifecycle() was not idempotent. It
assumed that each call to initLifecycle() would
eventually lead to a call to onAttach() in a 1:1
pairing. By ensuring that it only adds its
OnPreAttachedListener at most once, we make it
idempotent and fix a crash that occurs when
calling SavedStateRegistryController.performAttach()
multiple times.

Our original FragmentManagerTest that handles the
removal before onAttach() wasn't actually catching
that case consistently due to how the test thread
played with the main thread - that test was updated
in addition to using the same technique in the new
test.

BUG: 209514595
BUG: 209540551
Test: newly added FragmentManagerTest test
Change-Id: I27ff3e396774250ae036e07597f0028449759a8d