Pass context to GlanceAppWidget onDelete
onDelete is called outside of the Composable scope thus clients don't have access to the LocalContext.

In some cases the context might be needed, for example to fetch a dependency.

Since this method comes from the receiver that provides a context there is no harm on passing it.

Test: manual
Relnote: "Pass Context to GlanceAppWidget onDelete callback"

Change-Id: I4c7953e962072c4f88e0c34f6bff045c7fe3bbf1