Reduce the visual glitch after resuming the app.
- When app is resumed system uses the app screenshot to create the
illusion of responsiveness. However, this screenshot is automatically
removed shortly after the app is resumed and before we have a ready
surface to display our content. Resulting in a black screen being shown
right before the actual content is displayed.
This change takes the snapshot of the surface view when app is paused
and shows that behind the surface view. As a result users see the
snapshot instead of the black screen when resuming the app and before the
surface is recreated.

Test: Manual
Bug: 210686704
Relnote: Reduces the visual glitch on resume.
Change-Id: Iff7e07b38369d82620d5245e0d204d27eae6cfd0