Workaround for Red Mi Note 10 quirk
The symptom of the quirk is, when the PreviewView is in FILL_* mode, the scaled up preview is not cropped by its parent, thus covers the neighboring Views. Please find details in the bug.

The developer has verified that switching to COMPATIBLE mode fixed the issue. So in this CL, we force PreviewView to use TextureView if the quirk exists.

Bug: 211370840
Test: manual test and ./gradlew bOS
Change-Id: I5e8872558708bbb5396cccb03019de6d0bcdf26f