Fix PreviewProcessorTimestampTest's problem which doesn't work after the extensions related config was applied by the ExtensionEnabledCameraSelector
The original PreviewProcessorTimestampTest checked whether the PreviewConfig contains a PreviewCaptureProcessor then determined whether the test should be run. After the Extensions API design was changed that a extension enabled camera selector needs to be retrieved first, the PreviewCaptureProcessor will be added to the use case's config after the use case is bound. The original flow and logic doesn't work any more.

Instead, we need to provide a customized ExtendedCameraConfigProvider which allows the test to monitor and verify the frames' timestamps.

The original java implementation is removed and rewriting the test in kotlin.

Bug: 206924644
Test: PreviewProcessorTimestampTest
Change-Id: I38beba9029b9b988fa2c9a6253cd69503de80f3d