Ignore PreviewProcessorTimestampTest on specific OEM's devices which the timestamp is not kept
The test is to check whether the frame's timestamp is kpet the same after it is processed by the extension vendor library. So that the Preview's frames can be correctly mapped with ImageAnalysis' frames. Because this requirement is not described in the interface javadoc, OEMs might not implement it. Temporarily, this test will be ignored on the OEM's devices which the timestamp info is not kept.

Bug: 212499473
Test: PreviewProcessorTimestampTest
Change-Id: I2c3cdde41131f3dea3825505032366291702f1d4