Add @LabTestOnly on pipe-integration tests
The instrumental tests in the pipe-integration module
may still be flaky while we close the Surface.

It may request some feedback from the Camera-pipe library,
tests may not be able to be fixed in the short time, we will
temporarily disable this test in the black box testing and
only track in CameraX lab.
The further fix and update can be tracked in b/187015621

Bug: 208827848
Test: N/A
Change-Id: I6045df6081748ea4e2262d9237f535e5aebe69a2