Persist RemoteKey for V3RemoteMediator integration sample
Fix RemoteMediator from producing pages that are duplicate or out of
order in Paging's v3 sample app. This is a change I've been picking on
top of the sample whenever I need to test something - seems fit to just
merge this, since it is quite difficult to tell when changes to Paging
are working well without it as the same few items are repeated over and
over again.

Test: ./gradlew paging:integration-tests:testapp:assemble
Change-Id: Ic8fd141d432e51c69af542aa83558e4e8e5cc72b