Experimental Support for Hierarchical WF Styles
Note VersionedParcellable's limitations (you can't change the
serialization of a class in a list) meant we had to shoehorn
seralization of the Option's childSettings into the
UserStyleSetting instead of in the ListOption where it would have
been more natural.

Bug: 214504165
Test: Built a new sample and extended the existing sample editor and extended existing XML tests.
Relnote: We've added experimental support for hierarchical style schemas. We've added a new property to androidx.wear.watchface.style.UserStyleSetting.Option, childSettings which is initially only used by ListOption. This allows a hierarchy of styles to be described for use by Editor UIs, the underlying UserStyle is unchanged and is still a Map<String, ByteArray>.
Change-Id: Iaf6f4ba2862b8d1aff3ccad7f4b9bd2826b5db51