Bug fix with fillMaxSize/Width/Height in glance-wear-tiles
When assigning the top level container with fillMaxSize/Width/Heightmodifier in content composbale, nothing is displayed on the screen. The reason is we added a extra container Box as root during composition, which has default dimension modifier as "WRAP". A "EXPAND" Box/Column/Row in a "WRAP" root results in zero dimension.

Relnote: Bug fix with fillMaxSize/Width/Height in glance-wear-tiles
Test: sample demo
Change-Id: I0a39f94b75561d17410c6d164701625a2e61ad54