Remove CallingUID and callerHasSystemAccess from CallerAccess.
CallingUID and callerHasSystemAccess aren't really used in the local
backend.

We are going to create a new subclass CallerAccessAPI31 or
FrameworkCallerAccess in AppSearch framework, which will extends
CallerAccess and add AttributionSource.
The AttributionSource contains callingUID, so it doesn't make sense we
keep callingUID in CallerAccess and won't use it while it exists anyway
in it's subclass.

callerHasSystemAccess is not used in the local backend either, so move
it to the subclass seems a right move.


Bug: 215624105
Test: Presubmit
Change-Id: Ie036eab3f19f24b2c297168e000c1b5944c46448