Ensure constancy of watch face IDs
This patch combined with a system patch ensures consistency of
WatchState.watchFaceInstanceId and EditorSession.watchFaceId.

We only recently added support for id passed to headless and
we have decided to change the prefix used by the system to
signal compatibility.

Test: Manual testing
Bug: 218291962
Relnote: EditorSession.watchFaceId can now be used on all API levels. In addition its value will now always be consistent with WatchState.watchFaceInstanceId.
Change-Id: I323b949e624150437015cbe3bc522fba9fb4398a